pyproject.toml
[project]
name = "weather_bot"
version = "0.1.0"
description = "Script to fetch METAR/TAF, store in SQLite and send Telegram reports"
requires-python = ">=3.14"
readme = "README.md"
license = {text = "MIT"}

# Core run-time dependencies

dependencies = [
    "requests>=2.31",
    "python-metar>=1.4.0",
    "python-dateutil>=2.8",
    "tzdata>=2024.1"
]

[project.scripts]
weather-bot = "bot.cli:main"

[build-system]
requires = ["setuptools>=66", "wheel"]
build-backend = "setuptools.build_meta"
